djbot/pyproject.toml
[tool.poetry]
name = "djbot"
version = "0.1.0"
description = ""
authors = ["Anastasiia Maiorova <anastasiamayorova2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.16"
langchain-openai = "^0.1.4"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"
langchainhub = "^0.1.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "djbot.start_up:start_up"
